+++ GameLib/Mapa.h
@@ -7,7 +7,7 @@ protected:
 	sf::Vector2i foodPos;
 public: 
 	Map(sf::Vector2i size);
-	bool isPlayerOutside(sf::Vector2i pos, float radius);
+	bool isPlayerOutside(sf::Vector2f pos, float radius);
 
 	sf::Vector2i getSize();
 };
